Implemented IKEv1 attribute encoding in SA payload

+++ src/libcharon/encoding/payloads/transform_attribute.h
@@ -109,7 +109,7 @@ struct transform_attribute_t {
 	 *
 	 * @return 		value
 	 */
-	u_int16_t (*get_value) (transform_attribute_t *this);
+	u_int64_t (*get_value) (transform_attribute_t *this);
 
 	/**
 	 * Sets the value of the attribute.
